requirements.txt
streamlit==1.32.0
pandas==2.2.0
plotly==5.18.0
pillow==10.2.0
requests==2.31.0
google-cloud-storage==2.14.0
google-api-python-client==2.116.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
python-dotenv==1.0.1
google-generativeai==0.3.2
folium==0.15.1
streamlit-folium==0.18.0
snowflake-connector-python==3.7.0 